Changed new target field of user_id attribute from "created_by" to "updated_by"
Funny thing here: The old Dime API had a package included called "Blamable" which filled out the user_id on create of a new entity. But the old Dime frontend overwrote this ID with the user who updated the entity. So the blamable package of the old Dime did apparently nothing.